Refactor: move RaftState to dir raft_state
As `RaftState` grows in size and complexity while serving as the primary
state container in openraft, it is better to move it to a subfolder and
structure its subcomponents as separate files within it.
