+++ openraft/src/raft_state/vote_state_reader.rs
@@ -1,0 +1,8 @@
+use crate::NodeId;
+use crate::Vote;
+
+/// APIs to get vote.
+pub(crate) trait VoteStateReader<NID: NodeId> {
+    /// Get a reference to the current vote.
+    fn vote_ref(&self) -> &Vote<NID>;
+}
